% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el.R
\docType{data}
\name{el}
\alias{el}
\title{Winner-Loser Edgelist}
\format{A data frame with 2397 rows and 2 variables:
\describe{
  \item{winner}{The winner of an individual contest}
  \item{loser}{The loser of an individual contest}
}}
\usage{
el
}
\description{
A two variable dataframe with columns winner and loser
}
\keyword{datasets}

