% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/notexported.R
\name{funisi}
\alias{funisi}
\title{This function is used to calculate the I and SI given the dominance matrix}
\usage{
funisi(M)
}
\description{
This function is used to calculate the I and SI given the dominance matrix
}

