% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_wl_dataframe.R
\name{get_wl_dataframe}
\alias{get_wl_dataframe}
\title{Converts results dataframe to win-loss dataframe}
\usage{
get_wl_dataframe(df, ties = "remove")
}
\arguments{
\item{df}{A results dataframe}

\item{ties}{How to handle ties, default is \code{ties="remove"}
Alternative is \code{ties="keep"}}
}
\value{
A win-loss dataframe
}
\description{
Converts results dataframe to win-loss dataframe
}
\section{Further details}{

Add more detailed description.
Ensure that first 3 columns are id1, id2, result
don't have to be named that but must refer to those
Will keep all remaining columns intact
}
\examples{
get_wl_dataframe(randomtourney(8))
}

