% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wl_df.R
\name{get_wl_df}
\alias{get_wl_df}
\title{Converts results dataframe to win-loss dataframe}
\usage{
get_wl_df(df, ties = "remove")
}
\arguments{
\item{df}{A results dataframe}

\item{ties}{How to handle ties, default is \code{ties="remove"}
Alternative is \code{ties="keep"}}
}
\value{
A win-loss dataframe
}
\description{
Converts results dataframe to win-loss dataframe
}
\section{Further details}{

A results dataframe first 3 variables are id1, id2, result.
Results can be "W", "L", or "T" or "1", "0", "0.5".
The output will be a win-loss dataframe that will
reorganize the first 3 variables into winner, loser
and result (=1 for Win or =0.5 for ties).
}
\examples{
get_wl_df(randomtourney(8))
}

