% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isi13.R
\name{isi13}
\alias{isi13}
\title{Compute best ranked matrixed besed on I&SI method 2013}
\usage{
isi13(Mk, p = c(1, 0, 0, 0), a_max = 50, nTries = 30, p2)
}
\arguments{
\item{Mk}{A competition results or win-loss matrix}

\item{p}{A vector of probabilities for each of 4 methods}

\item{a_max}{Number of tries}

\item{nTries}{Number of iterations}

\item{p2}{probability for last method}
}
\value{
A computed ranked matrix best_matrix best_ranking I and SI
}
\description{
Compute best ranked matrixed besed on I&SI method 2013
}
\section{Further details}{

Add more detailed description.
See \code{\link{isi98}}: for further info.
}
\examples{
isi98(mouse)
}

