% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isi98.R
\name{isi98}
\alias{isi98}
\title{Calculate the rank order of a dominance matrix using De Vries' 1998 I&SI method}
\usage{
isi98(M, nTries)
}
\arguments{
\item{M}{A matrix with individuals ordered identically in rows and columns.}

\item{nTries}{Number of permutation tries.}
}
\value{
A list containing the most optimal matrix or matrices found, the
 rank order of individuals, the numbrer of inconsistencies \code{I} and the
 strength of inconsistencies \code{SI}
}
\description{
Calculate the rank order of a dominance matrix using De Vries' 1998 I&SI method
}
\section{References}{

Han de Vries (1998) de Vries H. 1998. Finding a dominance order most consistent
 with a linear hierarchy: a new procedure and review. Animal Behaviour 55:827-843
}

\section{Further details}{

none
}
\examples{
isi98(bonobos)
isi98(people)
}

