% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isi98b.R
\name{isi98b}
\alias{isi98b}
\title{Compute best ranked matrixed based on original I&SI method}
\usage{
isi98b(m, nTries = 100)
}
\arguments{
\item{m}{A  win-loss matrix}

\item{nTries}{Number of tries to find best order}
}
\value{
A computed ranked matrix best_matrix best_ranking I and SI
}
\description{
Compute best ranked matrixed based on original I&SI method
}
\section{Further details}{

Code based on algorithm described by de Vries, H. 1998. Finding a
dominance order most consistent with a linear hierarchy:
a new procedure and review. Animal Behaviour, 55, 827-843.
The number of iterations should be very high and/or the function
should be run several times to detect the optimal matrix or matrices.
 It may take several runs to find the matrix with the lowest SI,
 especially for very large matrices.
See \code{\link{isi13}}: for further info.
}
\examples{
isi98b(mouse,nTries=50)
}

