% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isi98.R
\name{matrix_change}
\alias{matrix_change}
\title{matrix_change' is used to modify the dominance matrix according to the give sequence}
\usage{
matrix_change(M, sequence)
}
\description{
matrix_change' is used to modify the dominance matrix according to the give sequence
}

