% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/notexported.R
\name{mtxbinom}
\alias{mtxbinom}
\title{Extra, not-exported, functions}
\usage{
mtxbinom(mtx, p = 0.5, alpha = 0.05)
}
\arguments{
\item{m}{A matrix with individuals ordered identically in rows and columns.}
}
\description{
Function required for transforming to 1/0 win/loss matrix based on binomial distribution
}

