% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isi13.R, R/notexported.R
\name{nature}
\alias{nature}
\title{nature' is used to offer the initial sequence depends on the porpotion of dominance}
\usage{
nature(M)

nature(M)
}
\description{
nature' is used to offer the initial sequence depends on the porpotion of dominance

nature' is used to offer the initial sequence depends on the porpotion of dominance
}

