% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\docType{data}
\name{people}
\alias{people}
\title{Human sociomatrix}
\format{A data frame with 6 rows and 6 variables:
\describe{
  \item{Adam}{The number of losses by human Adam against all other humans}
  \item{Bryan}{The number of losses by human Bryan against all other humans}
  \item{Chris}{The number of losses by human Chris against all other humans}
  \item{Derek}{The number of losses by human Derek against all other humans}
  \item{Eddie}{The number of losses by human Eddie against all other humans}
  \item{Frank}{The number of losses by human Frank against all other humans}
}}
\usage{
people
}
\description{
A win-loss sociomatrix with each row being the number of wins
against individual humans in each column.
}
\keyword{datasets}

