% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/return_wl_dataframe.R
\name{return_wl_dataframe}
\alias{return_wl_dataframe}
\title{Converts win-loss matrix of valued data to win-loss dataframe}
\usage{
return_wl_dataframe(mat)
}
\arguments{
\item{mat}{A win-loss matrix of valued data}
}
\value{
A win-loss dataframe
}
\description{
Converts win-loss matrix of valued data to win-loss dataframe
}
\section{Further details}{

The inputted matrix must be valued data with no ties.
The diagonal may be zeros or NA. The returned win-loss
dataframe will have three variables: winner, loser, result
 (which will all contain "W").
}
\examples{
return_wl_dataframe(bonobos)
}

