% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshps.R
\name{rshps}
\alias{rshps}
\title{Get Relaiontship Descriptives of Sociomatrix}
\usage{
rshps(m)
}
\arguments{
\item{m}{A matrix with individuals ordered identically in rows and columns.}
}
\value{
a list of total, unknown, tied, two-way and one-way relationships
}
\description{
Get Relaiontship Descriptives of Sociomatrix
}
\examples{
rshps(people)
}

