% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isi13.R
\name{shift}
\alias{shift}
\title{shift' function is used to put the rank ith individual at the jth position}
\usage{
shift(M, i, j)
}
\description{
shift' function is used to put the rank ith individual at the jth position
}

