% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseness.R
\name{sparseness}
\alias{sparseness}
\title{Calculate the sparseness of relationships in a sociomatrix}
\usage{
sparseness(m)
}
\arguments{
\item{m}{A matrix with individuals ordered identically in rows and columns.}
}
\value{
The sparseness of \code{m}
}
\description{
Calculate the sparseness of relationships in a sociomatrix
}
\section{Further details}{

The sparseness of a matrix is the proportion of null dyads
}
\examples{
sparseness(mouse)
sparseness(caribou)
}

