% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/notexported.R
\name{swap}
\alias{swap}
\title{This function is to modify the dominance matrix when ith individual and
jth individual exchange their positions.}
\usage{
swap(M, i, j)
}
\description{
This function is to modify the dominance matrix when ith individual and
jth individual exchange their positions.
}

