% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isi13.R
\name{transform}
\alias{transform}
\title{transform' function is the function to transform the original matrix to the fictitious matrix}
\usage{
transform(M)
}
\description{
transform' function is the function to transform the original matrix to the fictitious matrix
}

