% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttri_test.R
\name{ttri_test}
\alias{ttri_test}
\title{Signficance testing of Triangle Transitivity of Sociomatrix}
\usage{
ttri_test(m, Nperms = 1000)
}
\arguments{
\item{m}{A frequency or binary win-loss matrix}

\item{Nperms}{Number of randomizations in signficance test}
}
\value{
Pt, triangle transitivity and pvalue
}
\description{
Signficance testing of Triangle Transitivity of Sociomatrix
}
\section{Further details}{

Algorithm described in D. Shizuka and D. B. McDonald, 2012,
A social network perspective on transitivity and linearity
in dominance hierarchies.  Animal Behaviour.
DOI:10.1016/j.anbehav.2012.01.011
Code adapted from original code by Dai Shikua - see
http://www.shizukalab.com/toolkits/sna/triangle-transitivity
}
\examples{
ttri_test(mouse)
}

