% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknowns.R
\name{unknowns}
\alias{unknowns}
\title{Calculate the number of unknown relationships in a sociomatrix}
\usage{
unknowns(m)
}
\arguments{
\item{m}{A matrix with individuals ordered identically in rows and columns.}
}
\value{
The number of unknown relationships in \code{m}
}
\description{
Calculate the number of unknown relationships in a sociomatrix
}
\section{Further details}{

An unknown relationship is defined as one whereby M(i,j)==M(j,i)==0 .
The zeros in each cell of dyads that have this property may be
referred to as structural zeros.
}
\examples{
unknowns(mouse)
}

