% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlmat2df.R
\name{wlmat2df}
\alias{wlmat2df}
\title{Converts win-loss dataframe to win-loss matrix}
\usage{
wlmat2df(df, ties = "remove")
}
\arguments{
\item{df}{A results or win-loss dataframe}

\item{ties}{How to handle ties, default is \code{ties="remove"}
Alternative is \code{ties="keep"}}
}
\value{
A win-loss matrix
}
\description{
Converts win-loss dataframe to win-loss matrix
}
\section{Further details}{

Input dataframes or matrices with only 2 columns are
considered to be winners in column 1 and losers in column 2.
If input dataframe has three columns, the third column
will be the result of the interaction between column 1
subject and column 2 subject. The result can be in
the "W/L/T" format or "1/0/0.5" format.
See \code{\link{get_wl_df}}: for further info.
}
\examples{
wlmat2df(randomtourney(8))
wlmat2df(randomtourney(15,pties=.15))
wlmat2df(randomtourney(15,pties=.15),ties="keep")
wlmat2df(matrix(c(1,3,5,3,2,1,6,3,2,4,2,3,4,4),ncol=2))
}

