% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/devries.R
\name{devries}
\alias{devries}
\title{Calculate the linearity of a dominance hierarchy - De Vries' method}
\usage{
devries(m)
}
\arguments{
\item{m}{A matrix with individuals ordered identically in rows and columns.}
}
\value{
The modified Landau's h' value of \code{m}, associated p-value and
plot of randomization test distribution.
}
\description{
Calculate the linearity of a dominance hierarchy - De Vries' method
}
\section{References}{

Han de Vries (1995) An improved test of linearity in dominance hierarchies
containing unknown or tied relationships. Animal Behaviour 50 pp. 1375-1389.
}

\section{Further details}{

This code is an edited and improved version of code originally written by Dai Shizuka.
}

\section{Acknowledgements}{

Jochen Weber
}
\examples{
devries(bonobos)
}

