% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_wl_matrix.R
\name{get_wl_matrix}
\alias{get_wl_matrix}
\title{Converts results or win-loss dataframe to win-loss matrix}
\usage{
get_wl_matrix(df, ties = "remove")
}
\arguments{
\item{df}{A competition results or win-loss dataframe}

\item{ties}{How to handle ties, default is \code{ties="remove"}
Alternative is \code{ties="keep"}}
}
\value{
A win-loss matrix
}
\description{
Converts results or win-loss dataframe to win-loss matrix
}
\section{Further details}{

Add more detailed description.
See \code{\link{get_wl_dataframe}}: for further info.
}
\examples{
get_wl_matrix(randomtourney(8))
}

