% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isi13.R
\name{isi13}
\alias{isi13}
\title{Compute best ranked matrixed based on new I&SI method}
\usage{
isi13(m, p = c(1, 0, 0, 0), a_max = 50, nTries = 30, p2 = 0.5)
}
\arguments{
\item{m}{A win-loss matrix}

\item{p}{A vector of probabilities for each of 4 methods}

\item{a_max}{Number of tries}

\item{nTries}{Number of iterations}

\item{p2}{probability for last method}
}
\value{
A computed ranked matrix best_matrix best_ranking I and SI
}
\description{
Compute best ranked matrixed based on new I&SI method
}
\section{Further details}{

Code based on algorithm described by Schmid & de Vries 2013,
Finding a dominance order most consistent with a linear hierarchy:
An improved algorithm for the I&SI method, Animal Behaviour
86:1097-1105. This first implementation of this algorithm is not
very fast. The number of tries should be very high and/or the function
should be run several times to detect the optimal matrix or matrices.
It may take several runs to find the matrix with the lowest SI,
especially for very large matrices. For small matrices it may be more
efficient to use the older algorithm. See \code{\link{isi98}}:
for further info.
}
\examples{
isi13(people)
}

