% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isi98.R
\name{isi98}
\alias{isi98}
\title{Compute best ranked matrixed besed on I&SI method 1998}
\usage{
isi98(M, nTries)
}
\arguments{
\item{M}{A competition results or win-loss matrix}

\item{nTries}{Number of iterations}
}
\value{
A computed ranked matrix best_matrix best_ranking I and SI
}
\description{
Compute best ranked matrixed besed on I&SI method 1998
}
\section{Further details}{

Add more detailed description.
See \code{\link{isi98}}: for further info.
}
\examples{
isi98(mouse)
}

