% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isi98.R
\name{isi98}
\alias{isi98}
\title{Calculate the rank order of a matrix according to the original I&SI method}
\usage{
isi98(Mk, nTries)
}
\arguments{
\item{Mk}{A matrix with individuals ordered identically in rows and columns.}

\item{nTries}{The number of iterations}
}
\value{
The best ranked matrix \code{Mk}, the rank order of dominant individuals,
Imin - the minimum number of inconsistencies found in the best matrix,
SImin - the strength of inconsistencies found in the best matrix
}
\description{
Calculate the rank order of a matrix according to the original I&SI method
}
\section{References}{

Han de Vries (1998)
}

\section{Further details}{

Forthcoming.
}
\examples{
isi98(mouse)
}

